Create edwards_9.4
There are many developer tools out there and almost every browser has one.  In this article we will be looking at Firefox and Chrome developer tools.  There are so many tricks that you can do with both of these tools but we will only be focusing on three of them in this document.

“Chrome DevTools is a set of web developer tools built directly into the Google Chrome browser. DevTools can help you edit pages on-the-fly and diagnose problems quickly, which ultimately helps you build better websites, faster.
“  Chrome DevTools. Google, 10, February 2019, https://developers.google.com/web/tools/chrome-devtools/

Chrome might be the industry standard for web developer tools. One thing that you can do is manipulate the styling within the browser itself. Literally, assignments 9.2 and 9.3 were centralized around this topic. You can do this by finding the element that you’d like t inspect and either right clicking and “inspect” or by using fn F12 (on a mac). The developer tool panel will open and you can then select the element you are looking to modify. For the case of assignment 9.2 and 9.3 we were asked to modify the right and left columns by changing the background and font color.

Another trick that can be done right from the browser developer tool is to work directly on the DOM. “The DOM tree view in the Chrome DevTools Elements panel displays the DOM structure of the current web page. Live-edit the content and structure of your page through DOM updates. “Basques, Kayce; 29, January, 2019. Edit the DOM. https://developers.google.com/web/tools/chrome-devtools/inspect-styles/edit-dom. All in all the DOM , “is an interface that allows a programming language to manipulate the content, structure, and style of a website. JavaScript is the client-side scripting language that connects to the DOM in an internet browser. “ Rascia, Tania. 6, November, 2017. Introduction to the DOM. https://www.digitalocean.com/community/tutorials/introduction-to-the-dom
Why is this important? It’s important because while not typical, you could manipulate the html to render something different on the page. The typical use of the the DOM is to debug code.  One thing to note, while the changes you make working in the console render on screen, it is not a permanent fix and once refreshed all changes will be lost. 

Within Firefox, (and all browser developer tools) you have the page inspector, which will allow the user to modify and inspect the html and style of a page.  This is important because without this simple command, users wouldn’t be able to live edit on the page itself. By using fn f12 or right clicking a mouse and “inspect”  the developer tools pane would not open and debugging would be much more inefficient.

All in all, developer tools are an essential way developers can manipulate and debug code. There are so many options for one to select but they all have similar properties and abilities.  Some of the more basic features include, inspector, inspect and select colors, edit the css, edit fonts,  edit the html, repositioning elements on the page and  working with animations. Once you have a basic understanding of what is being viewed, working within the console is a much friendlier place. Happy debugging.